Thang/ai autoreply kb improvements

+++ deploy/cleanup.sql
@@ -9,7 +9,8 @@ BEGIN TRANSACTION;
 DELETE FROM messages;
 DELETE FROM conversation_labels;
 DELETE FROM conversation_notes;
-DELETE FROM conversation_read_state;
+IF OBJECT_ID(N'dbo.conversation_read_state', N'U') IS NOT NULL
+    DELETE FROM conversation_read_state;
 DELETE FROM conversations;
 DELETE FROM processed_messages;
 DELETE FROM contact_external_ids;
